requirements.txt
amplitude_analytics~=1.1.1
dataclasses-json~=0.6.7
sseclient-py~=1.8.0
